pull Civet webserver into prometheus.{hpp,cpp}, put in opaque PrometheusServer class
